Optimize the evaluation of applicability
The goal is to cut some computations that are useless to determine if a
rule is applicable, such as computing the terms of a sum. Since less
rules are used during the computation, we also have less
“missingVariables” witch is desired.

The implementation is a bit rough, it would probably be cleaner to have
a separate `evaluateApplicability` method on the engine instance, but
it's a bit difficult to re-implement it for all mechanisms, we need a
subset of base mechanisms before.